move back into pkg component of games-commons
