+++ gazeplay-games/src/main/java/net/gazeplay/games/colors/ColorBox.java
@@ -16,6 +16,7 @@ import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import net.gazeplay.IGameContext;
 import net.gazeplay.commons.gaze.devicemanager.GazeEvent;
+import net.gazeplay.components.AbstractGazeIndicator;
 
 @Slf4j
 public class ColorBox extends StackPane {
